feat/order: long -> Interger 수정 및 임시 값 추가

+++ src/main/java/org/example/eatopia/domain/order/dto/response/OrderDetailProductResponse.java
@@ -10,7 +10,7 @@ public record OrderDetailProductResponse(
         Long productId,
         String productName,
         Long sellerId,
-        Long quantity,
+        Integer quantity,
         BigDecimal price
 ) {
     public static OrderDetailProductResponse from(final OrderDetail orderDetail) {
